add error handling when allocation failed

+++ shm/socket.c
@@ -12,7 +12,7 @@
 
 #define MAX_PID_BITS (22)
 #define GET_LOCAL_PORT_FROM_FD(x) ((x)&0xFFFF)
-#define GET_REMOTE_PORT_FROM_FD(x) ((x) >> 16)
+#define GET_REMOTE_PORT_FROM_FD(x) (((x) >> 16) & 0xFFFF)
 #define FD2PORT(fd) (((fd) << MAX_PID_BITS) | getpid())
 
 #define LISTEN_PID_TABLE_KEY (0x4f4f4f4f)
@@ -73,6 +73,7 @@ int get_free_port(int idx) {
 //int socket(int domain, int type, int protocol) {}
 
 int bind(int fd, const struct sockaddr *addr, socklen_t len) {
+	//printf("bind: \n");
 	struct sockaddr_in *addr_in = (struct sockaddr_in *)addr;
 	int port = addr_in->sin_port;
 	if (0 != read_port_table(port)) {
@@ -114,7 +115,9 @@ int connect(int fd, const struct sockaddr *addr, socklen_t len) {
 		return -1;
 	}
 	client_fd_to_idx[fd] = idx;
-	rb_get(get_idx(GET_REMOTE_PORT_FROM_FD(idx), GET_LOCAL_PORT_FROM_FD(idx)));
+	if (!rb_get(get_idx(GET_REMOTE_PORT_FROM_FD(idx), GET_LOCAL_PORT_FROM_FD(idx)))) {
+		return -3;
+	}
 	// send signal to pid
 	kill(pid, SIGUSR1);
 	return 0;
@@ -136,7 +139,9 @@ int accept(int fd, struct sockaddr *addr, socklen_t *len) {
 	sigaction(SIGUSR1, &siga, NULL);
 	pause();
 	// wait for signal
-	rb_get(get_idx(GET_REMOTE_PORT_FROM_FD(fd), GET_LOCAL_PORT_FROM_FD(fd)));
+	if (!rb_get(get_idx(GET_REMOTE_PORT_FROM_FD(fd), GET_LOCAL_PORT_FROM_FD(fd)))) {
+		return -1;
+	}
 	int client_port = read_port_table(server_port) & 0xFFFF;
 	return (client_port << 16) | server_port;
 }
@@ -152,8 +157,9 @@ ssize_t sendto(int fd, const void *buf, size_t len, int flags,
 		fd = client_fd_to_idx[fd];
 	}
 	writer = rb_get(get_idx(GET_LOCAL_PORT_FROM_FD(fd), GET_REMOTE_PORT_FROM_FD(fd)));
+	if (!writer) return -1;
 #ifndef NDEBUG
-	printf("socket send API: writer->index: %d, writer: %x, address: %x", writer->index, writer, writer->address);
+	printf("socket send API: writer->index: %d, writer: %x, address: %x\n", writer->index, writer, writer->address);
 #endif	
 	return rb_write(writer, len, (char *)buf);
 }
@@ -171,6 +177,7 @@ ssize_t recvfrom(int fd, void *buf, size_t len, int flags,
 		fd = client_fd_to_idx[fd];
 	}
 	reader = rb_get(get_idx(GET_REMOTE_PORT_FROM_FD(fd), GET_LOCAL_PORT_FROM_FD(fd)));
+	if (!reader) return -1;
 #ifndef NDEBUG
 	printf("socket recv API: reader->index: %d, reader: %x, address: %x\n", reader->index, reader, reader->address);
 #endif
@@ -192,6 +199,7 @@ int close(int fd) {
 	write_port_table(GET_LOCAL_PORT_FROM_FD(fd), 0);
 	RingBuffer_t *reader = NULL;
 	reader = rb_get(get_idx(GET_REMOTE_PORT_FROM_FD(fd), GET_LOCAL_PORT_FROM_FD(fd)));
+	if (!reader) return -1;
 #ifndef NDEBUG
 	printf("socket close API: fd = %d, src: %d, dst: %d, reader->index: %d\n", fd, GET_REMOTE_PORT_FROM_FD(fd), GET_LOCAL_PORT_FROM_FD(fd), reader->index);
 #endif
